feat: Added Bio view to contact cards
