+++ package.json
@@ -39,6 +39,7 @@
 	"dependencies": {
 		"@babel/runtime": "^7.11.2",
 		"@material-ui/core": "^4.11.0",
+		"@material-ui/lab": "^4.0.0-alpha.56",
 		"@reduxjs/toolkit": "^1.3.6",
 		"@widgetbot/react-embed": "^1.2.2",
 		"aws-lambda": "^1.0.6",
@@ -59,7 +60,7 @@
 		"react-list": "^0.8.15",
 		"react-markdown": "^4.3.1",
 		"react-redux": "^7.2.0",
-		"react-scripts": "^3.4.1",
+		"react-scripts": "^3.4.3",
 		"react-vertical-timeline-component": "^3.0.2",
 		"redux": "^4.0.5",
 		"styled-components": "^4.1.3",
